8228336: Refactor native library loading implementation Reviewed-by: alanb, dholmes
