+++ src/java.base/share/classes/java/lang/System.java
@@ -2276,11 +2276,6 @@ public final class System {
             public void setCause(Throwable t, Throwable cause) {
                 t.setCause(cause);
             }
-
-            public void loadLibrary(Class<?> caller, String library) {
-                assert library.indexOf(java.io.File.separatorChar) < 0;
-                ClassLoader.loadLibrary(caller, library, false);
-            }
         });
     }
 }
