+++ src/java.base/windows/classes/jdk/internal/loader/ClassLoaderHelper.java
@@ -23,7 +23,7 @@
  * questions.
  */
 
-package java.lang;
+package jdk.internal.loader;
 
 import java.io.File;
